+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Social Share Change Log
 2.3.0 Under development
 -----------------------
 * Enh #16: Adds `enableIcons` and `enableDefaultAsset` option to the default configurator instead of `enableDefaultIcons` (greeflas, dimmitri)
+* Enh: Improves architecture of configurator, adds more abstraction to reduce dependency to the default implementation in widget (greeflas)
 * Chg: Deprecate `enableDefaultIcons` option of the default configurator (greeflas)
 
 2.2.0 April 22, 2018